[learn-squeryl] id is now an 'int'.

+++ learn_scala/squeryl/src/main/scala/pkg/db/T1.scala
@@ -1,6 +1,6 @@
 package pkg.db
 
-case class T1(id: Long,
+case class T1(id: Int,
               ver: Int,
               name: String,
               nicName: Option[String]) extends HasIdVer
